requirements.txt
streamlit
openai
psycopg2
pandas
python-dotenv
matplotlib
plotly
